fixed the controller to integrate with front end

+++ Business/Interfaces/ITransactionService.cs
@@ -1,11 +1,11 @@
 ﻿    public interface ITransactionService
     {
 
-     Task<OperationResult<bool>> TransferToAccountAsync(string senderAccountNumber,string recieverAccountNumber,int amount);
+     Task<OperationResult<bool>> TransferToAccountAsync(string senderAccountNumber,string recieverAccountNumber,decimal amount);
       
-     Task<OperationResult<bool>> DepositToAccountAsync(string accountNumber,int amount);
+     Task<OperationResult<bool>> DepositToAccountAsync(string accountNumber,decimal amount);
 
-     Task<OperationResult<bool>> WithdrawFromAccountAsync(string accountNumber,int amount);
+     Task<OperationResult<bool>> WithdrawFromAccountAsync(string accountNumber,decimal amount);
             
     }
